+++ brokers/test/utils.go
@@ -10,7 +10,7 @@
 //   Red Hat, Inc. - initial API and implementation
 //
 
-package brokers_test
+package test
 
 import (
 	"bytes"
